cleared a compiler err
git-svn-id: svn://orca.physics.unc.edu/Orca64@11943 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Custom Hardware/SNO+/hiredis/dict.c
@@ -98,7 +98,7 @@ static int _dictInit(dict *ht, dictType *type, void *privDataPtr) {
 }
 
 /* Expand or create the hashtable */
-static int dictExpand(dict *ht, unsigned long size) {
+static int dictExpand(dict *ht, uint32_t size) {
     dict n; /* the new hashtable */
     unsigned long realsize = _dictNextPower(size), i;
